feature: Add plugins system
The project code is spread across multiple workspaces.

Added mechanism for dynamic loading of dynamic libraries for connecting code execution plugins.

Due to difficulties with maintaining and managing dependencies in an
up-to-date state, the deno runtime was extracted into a separate
plugin that is dynamically connected to the devrcfile config.